+++ spec/controllers/general_controller_spec.rb
@@ -410,5 +410,4 @@ describe GeneralController, 'when using xapian search' do
     end
 
   end
-
 end