+++ spec/models/user_spec.rb
@@ -442,6 +442,73 @@ describe User do
 
   end
 
+  describe '.all_time_requesters' do
+    it 'gets most frequent requesters' do
+      # FIXME: This uses fixtures. Change it to use factories when we can.
+      expect(User.all_time_requesters).to eql(
+        {
+          users(:bob_smith_user) => 5,
+          users(:robin_user) => 2,
+          users(:another_user) => 1
+        }
+      )
+    end
+  end
+
+  describe '.last_28_day_requesters' do
+    it 'gets recent frequent requesters' do
+      user_with_3_requests = FactoryGirl.create(:user)
+      3.times { FactoryGirl.create(:info_request, user: user_with_3_requests) }
+      user_with_2_requests = FactoryGirl.create(:user)
+      2.times { FactoryGirl.create(:info_request, user: user_with_2_requests) }
+      user_with_1_request = FactoryGirl.create(:user)
+      FactoryGirl.create(:info_request, user: user_with_1_request)
+      user_with_an_old_request = FactoryGirl.create(:user)
+      FactoryGirl.create(:info_request, user: user_with_an_old_request, created_at: 2.months.ago)
+
+      expect(User.last_28_day_requesters).to eql(
+        {
+          user_with_3_requests => 3,
+          user_with_2_requests => 2,
+          user_with_1_request => 1
+        }
+      )
+    end
+  end
+
+  describe '.all_time_commenters' do
+    it 'gets most frequent commenters' do
+      # FIXME: This uses fixtures. Change it to use factories when we can.
+      expect(User.all_time_commenters).to eql(
+        {
+          users(:bob_smith_user) => 1,
+          users(:silly_name_user) => 1
+        }
+      )
+    end
+  end
+
+  describe '.last_28_day_commenters' do
+    it 'gets recent frequent commenters' do
+      user_with_3_comments = FactoryGirl.create(:user)
+      3.times { FactoryGirl.create(:comment, user: user_with_3_comments) }
+      user_with_2_comments = FactoryGirl.create(:user)
+      2.times { FactoryGirl.create(:comment, user: user_with_2_comments) }
+      user_with_1_comment = FactoryGirl.create(:user)
+      FactoryGirl.create(:comment, user: user_with_1_comment)
+      user_with_an_old_comment = FactoryGirl.create(:user)
+      FactoryGirl.create(:comment, user: user_with_an_old_comment, created_at: 2.months.ago)
+
+      expect(User.last_28_day_commenters).to eql(
+        {
+          user_with_3_comments => 3,
+          user_with_2_comments => 2,
+          user_with_1_comment => 1
+        }
+      )
+    end
+  end
+
   describe '#transactions' do
 
     it 'returns a TransactionCalculator with the default transaction set' do